# openapi/oh.yaml
openapi: 3.0.1
info:
  title: Open Hospital API Documentation
  description: Open Hospital API Documentation
  contact:
    name: ApiInfo.DEFAULT_CONTACT
  license:
    name: GPL-3.0 license
    url: https://github.com/informatici/openhospital-api?tab=GPL-3.0-1-ov-file#readme
  version: 0.1.0
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: [ ]
paths:
...
  /malnutritions:
    put:
      tags:
        - Malnutritions
      operationId: updateMalnutrition
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MalnutritionDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MalnutritionDTO"
      security:
        - bearerAuth: [ ]
    post:
      tags:
        - Malnutritions
      operationId: newMalnutrition
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MalnutritionDTO"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MalnutritionDTO"
      security:
        - bearerAuth: [ ]
    delete:
      tags:
        - Malnutritions
      operationId: deleteMalnutrition
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: boolean
      security:
        - bearerAuth: [ ]
  /laboratories/{code}:
    get:
      tags:
        - Laboratories
      operationId: getExamById
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LaboratoryDTO"
      security:
        - bearerAuth: [ ]
    put:
      tags:
        - Laboratories
      operationId: updateLaboratory
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabWithRowsDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: boolean
      security:
        - bearerAuth: [ ]
    delete:
      tags:
        - Laboratories
      operationId: deleteExam
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: boolean
      security:
        - bearerAuth: [ ]
...
  /exams/{code}:
    put:
      tags:
        - Exams
      operationId: updateExam
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamWithRowsDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamDTO"
      security:
        - bearerAuth: [ ]
    delete:
      tags:
        - Exams
      operationId: deleteExam_1
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: boolean
      security:
        - bearerAuth: [ ]
...
  /exams:
    get:
      tags:
        - Exams
      operationId: getExams
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExamDTO"
      security:
        - bearerAuth: [ ]
    post:
      tags:
        - Exams
      operationId: newExam
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamWithRowsDTO"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamDTO"
      security:
        - bearerAuth: [ ]
...
  /malnutritions/{id_admission}:
    get:
      tags:
        - Malnutritions
      operationId: getMalnutrition
      parameters:
        - name: id_admission
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MalnutritionDTO"
      security:
        - bearerAuth: [ ]
  /malnutritions/last/{id_patient}:
    get:
      tags:
        - Malnutritions
      operationId: getLastMalnutrition
      parameters:
        - name: id_patient
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MalnutritionDTO"
      security:
        - bearerAuth: [ ]
...
  /exams/description/{description}:
    get:
      tags:
        - Exams
      operationId: getExams_1
      parameters:
        - name: description
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExamDTO"
      security:
        - bearerAuth: [ ]
  /examrows/{code}:
    get:
      tags:
        - Exam Rows
      operationId: getExamRowsByCode
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExamRowDTO"
      security:
        - bearerAuth: [ ]
    delete:
      tags:
        - Exam Rows
      operationId: deleteExam_2
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: boolean
      security:
        - bearerAuth: [ ]
...
components:
  schemas:
    WardDTO:
      required:
        - beds
        - description
        - docs
        - nurs
        - visitDuration
      type: object
      properties:
        code:
          maxLength: 3
          type: string
          description: The ward code
        description:
          maxLength: 50
          type: string
          description: The name of the ward
        telephone:
          maxLength: 50
          type: string
          description: The phone number of the ward
        fax:
          maxLength: 50
          type: string
          description: The fax number of the ward
        email:
          maxLength: 50
          type: string
          description: The email of the ward
        beds:
          type: integer
          format: int32
        nurs:
          type: integer
          format: int32
        docs:
          type: integer
          format: int32
        visitDuration:
          type: integer
          format: int32
        lock:
          type: integer
          description: lock
          format: int32
          example: 0
        opd:
          type: boolean
        female:
          type: boolean
        male:
          type: boolean
        pharmacy:
          type: boolean
    PatientDTO:
      required:
        - age
        - bloodType
        - city
        - fatherName
        - firstName
        - motherName
        - secondName
        - sex
      type: object
      properties:
        code:
          type: integer
          format: int32
          readOnly: true
        firstName:
          maxLength: 50
          type: string
          description: First name of the patient
          example: Mario
        secondName:
          maxLength: 50
          type: string
          description: Last name of the patient
          example: Rossi
        name:
          type: string
          readOnly: true
        birthDate:
          type: string
          description: Birth date
          example: 1979-05-01
        age:
          type: integer
          description: Age
          format: int32
          example: 40
        agetype:
          type: string
          description: Age type
          example: "null"
        sex:
          type: string
          description: Sex
          example: M
          enum:
            - M
            - F
        address:
          maxLength: 50
          type: string
          description: Address
          example: "Via Roma, 12"
        city:
          maxLength: 50
          type: string
          description: City
          example: Verona
        nextKin:
          type: string
          readOnly: true
        telephone:
          maxLength: 50
          type: string
          description: Telephone
          example: "+393456789012"
        note:
          maxLength: 65535
          type: string
          description: Note
          example: Test insert new patient
        motherName:
          maxLength: 50
          type: string
          description: Mother's name
          example: Roberta
        mother:
          type: string
          description: "Mother's status (D=dead, A=alive)"
          example: A
          enum:
            - D
            - A
        fatherName:
          maxLength: 50
          type: string
          description: Father's name
          example: Giuseppe
        father:
          type: string
          description: "Father's status (D=dead, A=alive)"
          example: D
          enum:
            - D
            - A
        bloodType:
          type: string
          description: "Blood type (0-/+, A-/+ , B-/+, AB-/+)"
          example: A+
          enum:
            - 0-
            - 0+
            - A-
            - A+
            - B-
            - B+
            - AB-
            - AB+
        hasInsurance:
          type: string
          description: "HasInsurance (Y=Yes, N=no)"
          example: "N"
          enum:
            - "Y"
            - "N"
        parentTogether:
          type: string
          description: "Parent together (Y=Yes, N=no)"
          example: "N"
          enum:
            - "Y"
            - "N"
        taxCode:
          maxLength: 30
          type: string
          description: Tax code
          example: RSSMRA79E01L781N
        lock:
          type: integer
          description: Lock
          format: int32
          example: 0
        blobPhoto:
          type: array
          description: BlobPhoto
          items:
            type: string
            description: BlobPhoto
            format: byte
        hashCode:
          type: integer
          format: int32
          readOnly: true
        allergies:
          maxLength: 255
          type: string
          description: Allergies of patient
        anamnesis:
          maxLength: 255
          type: string
          description: Current anamnesis
        status:
          type: string
          description: Status
          example: I
          enum:
            - I
            - O
        consensusFlag:
          type: boolean
          description: Consensus flag
          example: true
        consensusServiceFlag:
          type: boolean
          description: Consensus service flag
          example: true
      description: Class representing a patient
...
    AdmissionDTO:
      required:
        - admDate
        - admitted
        - deleted
        - type
      type: object
      properties:
        id:
          type: integer
          description: Admission key
          format: int32
          example: 12
        admitted:
          type: integer
          description: If admitted or not
          format: int32
          example: 0
        type:
          type: string
          description: Type of admission
          example: "N"
        ward:
          $ref: "#/components/schemas/WardDTO"
        patient:
          $ref: "#/components/schemas/PatientDTO"
        admDate:
          type: string
          description: Admission date
        admType:
          $ref: "#/components/schemas/AdmissionTypeDTO"
        diseaseIn:
          $ref: "#/components/schemas/DiseaseDTO"
        diseaseOut1:
          $ref: "#/components/schemas/DiseaseDTO"
        diseaseOut2:
          $ref: "#/components/schemas/DiseaseDTO"
        diseaseOut3:
          $ref: "#/components/schemas/DiseaseDTO"
        operation:
          $ref: "#/components/schemas/OperationDTO"
        opDate:
          type: string
          description: Operation date
        opResult:
          type: string
          description: 'Operation result value is ''P'' or ''N'' '
          example: "N"
        disDate:
          type: string
          description: Discharge date
        disType:
          $ref: "#/components/schemas/DischargeTypeDTO"
        note:
          maxLength: 65535
          type: string
          description: Free note
        transUnit:
          type: number
          description: Transfusional unit
          format: float
        visitDate:
          type: string
          description: Visit date
        pregTreatmentType:
          $ref: "#/components/schemas/PregnantTreatmentTypeDTO"
        deliveryDate:
          type: string
          description: Delivery date
        deliveryType:
          $ref: "#/components/schemas/DeliveryTypeDTO"
        deliveryResult:
          $ref: "#/components/schemas/DeliveryResultTypeDTO"
        weight:
          type: number
          description: Weight
          format: float
        ctrlDate1:
          type: string
          format: date-time
        ctrlDate2:
          type: string
          format: date-time
        abortDate:
          type: string
          format: date-time
        userID:
          type: string
          description: User id
        hashCode:
          type: integer
          format: int32
          readOnly: true
        lock:
          type: integer
          description: Lock
          format: int32
          example: 0
        deleted:
          type: string
          description: "Flag record deleted, values are 'Y' OR 'N' "
          example: "N"
        yprog:
          type: integer
          format: int32
        fhu:
          type: string
      description: The admission
    AdmissionTypeDTO:
      required:
        - code
        - description
      type: object
      properties:
        code:
          type: string
          description: Code of the admission type
          example: A
        description:
          maxLength: 50
          type: string
          description: Description of the admission type
          example: AMBULANCE
      description: Admission type
...
    DischargeTypeDTO:
      required:
        - code
        - description
      type: object
      properties:
        code:
          type: string
          description: Code of the discharge type
          example: SN
        description:
          maxLength: 50
          type: string
          description: Description of the discharge type
          example: SORTIE NORMALE
      description: DisChargeType
    DiseaseDTO:
      required:
        - code
        - description
        - diseaseType
        - ipdInInclude
        - ipdOutInclude
        - opdInclude
      type: object
      properties:
        code:
          maxLength: 10
          type: string
          description: Disease code
          example: "99"
        description:
          maxLength: 50
          type: string
          description: Disease description
        diseaseType:
          $ref: "#/components/schemas/DiseaseTypeDTO"
        opdInclude:
          type: boolean
          description: Indicates whether the disease is an OPD disease
          example: true
        ipdInInclude:
          type: boolean
          description: Indicates whether the disease is an IPD-IN disease
          example: true
        ipdOutInclude:
          type: boolean
          description: Indicates whether the disease is an IPD-OUT disease
          example: true
        hashCode:
          type: integer
          format: int32
          readOnly: true
        lock:
          type: integer
          description: Lock
          format: int32
          example: 0
      description: Class representing a disease
...
    MalnutritionDTO:
      required:
        - admission
        - dateSupp
        - height
        - weight
      type: object
      properties:
        code:
          type: integer
          description: The code malnutrition control
          format: int32
          example: 1
        dateSupp:
          type: string
          description: The date of this malnutrition control
          example: 1979-05-01T11:20:33
        dateConf:
          type: string
          description: The date of the next malnutrition control
          example: 1979-05-01T11:20:33
        admission:
          $ref: "#/components/schemas/AdmissionDTO"
        height:
          type: number
          description: The height of the patient
          format: float
          example: 165
        weight:
          type: number
          description: The weight of the patient
          format: float
          example: 65
        lock:
          type: integer
          description: Lock
          format: int32
          example: 0
    ExamDTO:
      type: object
      properties:
        code:
          maxLength: 10
          type: string
          description: Exam Code
          example: "01.01"
        description:
          maxLength: 100
          type: string
          description: Exam Description
          example: 1.1 HB
        procedure:
          type: integer
          description: Exam Procedure
          format: int32
          example: 1
        defaultResult:
          maxLength: 50
          type: string
          description: Exam Default Result
          example: '>=12 (NORMAL)'
        examtype:
          $ref: "#/components/schemas/ExamTypeDTO"
        lock:
          type: integer
          description: Lock
          format: int32
          example: 0
      description: Laboratory Exam
    ExamTypeDTO:
      type: object
      properties:
        code:
          type: string
          description: Exam Type Code
          example: HB
        description:
          maxLength: 50
          type: string
          description: Exam Type Description
          example: 1.Haematology
      description: Exam Type
    LabWithRowsDTO:
      type: object
      properties:
        laboratoryDTO:
          $ref: "#/components/schemas/LaboratoryDTO"
        laboratoryRowList:
          type: array
          items:
            type: string
    LaboratoryDTO:
      type: object
      properties:
        code:
          type: integer
          description: Laboratory Code
          format: int32
        material:
          type: string
          description: Laboratory Material
          example: Blood
        exam:
          $ref: "#/components/schemas/ExamDTO"
        registrationDate:
          type: string
          description: Laboratory Registration Date
        labDate:
          type: string
          description: Laboratory Exam Date
        result:
          type: string
          description: Laboratory Result
        lock:
          type: integer
          description: Lock
          format: int32
          example: 0
        note:
          maxLength: 255
          type: string
          description: Laboratory Note
          example: Note by laboratorist
        patientCode:
          type: integer
          description: Laboratory Patient Code
          format: int32
        patName:
          type: string
          description: Laboratory Patient Name
        inOutPatient:
          type: string
          description: Laboratory Patient InOut
          example: "0"
          enum:
            - I
            - O
        age:
          type: integer
          description: Laboratory Patient Age
          format: int32
        sex:
          type: string
          description: Laboratory Patient Sex
          example: M
        status:
          type: string
          description: Laboratory status
          example: DRAFT
          enum:
            - draft
            - open
            - done
            - invalid
            - deleted
...
    ExamWithRowsDTO:
      required:
        - exam
      type: object
      properties:
        exam:
          $ref: "#/components/schemas/ExamDTO"
        rows:
          type: array
          description: Possible result for the exam(only for exams with procedure
            1 and 2)
          example: "['>=12 (NORMAL)', 'IRREGULAR']"
          items:
            type: string
            description: Possible result for the exam(only for exams with procedure
              1 and 2)
            example: "['>=12 (NORMAL)', 'IRREGULAR']"
...
    ExamRowDTO:
      type: object
      properties:
        code:
          type: integer
          description: Exam Row Code
          format: int32
          example: 999
        description:
          type: string
          description: Exam Row Code
          example: NEGATIVE
        exam:
          $ref: "#/components/schemas/ExamDTO"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT